chore(base): move readonly db context to impelmentations
