+++ src/Christofel.Api/Ctu/Resolvers/DuplicateResolver.cs
@@ -6,6 +6,7 @@ using Christofel.Api.Ctu.Auth;
 using Christofel.BaseLib.Database;
 using Christofel.BaseLib.Database.Models;
 using Christofel.BaseLib.Extensions;
+using Christofel.BaseLib.Implementations.ReadOnlyDatabase;
 using Christofel.BaseLib.User;
 using Microsoft.EntityFrameworkCore;
