+++ src/Christofel.Application/Permissions/DbPermissionsResolver.cs
@@ -6,6 +6,7 @@ using Christofel.BaseLib.Database;
 using Christofel.BaseLib.Database.Models;
 using Christofel.BaseLib.Database.Models.Enums;
 using Christofel.BaseLib.Extensions;
+using Christofel.BaseLib.Implementations.ReadOnlyDatabase;
 using Christofel.BaseLib.Permissions;
 using Microsoft.EntityFrameworkCore;
