+++ src/Christofel.Management/CtuUtils/CtuIdentityResolver.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Christofel.BaseLib.Database;
 using Christofel.BaseLib.Database.Models;
 using Christofel.BaseLib.Extensions;
+using Christofel.BaseLib.Implementations.ReadOnlyDatabase;
 using Christofel.BaseLib.User;
 using Microsoft.EntityFrameworkCore;
 using Remora.Discord.Core;
